refactor: mv component
