+++ src/renderer/shared/lib/utils/index.ts
@@ -17,4 +17,3 @@ export * from './search';
 export * from './step';
 export * from './createQueuedRequest';
 export * from './createAsyncTaskPool';
-export { AsyncItem } from './AsyncItem';
