+++ src/renderer/shared/ui-kit/index.ts
@@ -7,6 +7,7 @@ export { SearchInput } from './SearchInput/SearchInput';
 export { ThemeProvider } from './Theme/ThemeProvider';
 export { ScrollArea } from './ScrollArea/ScrollArea';
 export { Accordion } from './Accordion/Accordion';
+export { AsyncItem } from './AsyncItem/AsyncItem';
 export { CardStack } from './CardStack/CardStack';
 export { InputFile } from './InputFile/InputFile';
 export { Progress } from './Progress/Progress';
